Fix bug 6126 (Shouldn't be able to delete events from read-only collections), r=grant: Make sure we CanDelete before calling onDeleteEvent directly (and touch up an assert message).
git-svn-id: http://svn.osafoundation.org/chandler/trunk@11058 3c49585b-f0f7-0310-b5f9-dfe92a88fbfe

+++ chandler/parcels/osaf/framework/blocks/calendar/CollectionCanvas.py
@@ -783,8 +783,11 @@ class wxCollectionCanvas(DragAndDrop.DropReceiveWidget,
         
         if ('__WXMAC__' in wx.PlatformInfo and 
             (keyCode in (wx.WXK_BACK, wx.WXK_DELETE))):
-            self.blockItem.onDeleteEvent(event)
-            return
+            # Fix bug 6126: make sure we can delete before calling
+            # onDeleteEvent (and eat the key event either way).
+            if self.blockItem.CanDelete():
+                self.blockItem.onDeleteEvent(event)
+            return # eat the key event.
 
         # didn't handle it so propagate the key event upward
         event.Skip()
